[env] Re-structure module tree
