--- src/engine/mod.rs
@@ -1,3 +1,0 @@
-pub(crate) mod game;
-pub(crate) mod loader;
-pub(crate) mod renderer;
