+++ src/game/engine/loader.rs
@@ -1,7 +1,11 @@
-use crate::browser;
-
+use crate::game::engine::browser::accessor::{LoopClosure, now, request_animation_frame};
+use crate::game::engine::browser::context::context;
+use crate::game::engine::browser::wrapper::create_raf_closure;
+use crate::game::engine::renderer::Renderer;
+use crate::game::engine::{Game, GameLoop, browser};
 use anyhow::{Error, anyhow};
 use serde_wasm_bindgen::from_value;
+use std::cell::RefCell;
 use std::rc::Rc;
 use std::sync::Mutex;
 use wasm_bindgen::closure::Closure;
