+++ src/game/engine/renderer.rs
@@ -1,7 +1,12 @@
 use anyhow::{Error, anyhow};
 use web_sys::{CanvasRenderingContext2d, HtmlImageElement};
 
-use crate::graphics::rect::Rect;
+pub(crate) struct Rect {
+  pub x: f32,
+  pub y: f32,
+  pub width: f32,
+  pub height: f32,
+}
 
 pub(crate) struct Renderer {
   context: CanvasRenderingContext2d,
