--- src/graphics/mod.rs
@@ -1,3 +1,0 @@
-pub mod color;
-pub mod point;
-pub mod rect;
