--- src/graphics/point.rs
@@ -1,10 +1,0 @@
-pub struct Point {
-  pub x: f64,
-  pub y: f64,
-}
-
-impl Point {
-  pub fn new(x: f64, y: f64) -> Self {
-    Self { x, y }
-  }
-}
