--- src/graphics/rect.rs
@@ -1,6 +1,0 @@
-pub(crate) struct Rect {
-  pub x: f32,
-  pub y: f32,
-  pub width: f32,
-  pub height: f32,
-}
