+++ src/main.rs
@@ -1,12 +1,9 @@
-mod browser;
-mod engine;
-mod game;
-mod graphics;
-
-use crate::browser::wrapper::spawn_local;
-use crate::engine::game::GameLoop;
-use crate::game::WalkTheDog;
 use wasm_bindgen::{JsCast, JsValue};
+
+mod game;
+use game::WalkTheDog;
+use game::engine::GameLoop;
+use game::engine::browser::wrapper::spawn_local;
 
 pub fn main() -> Result<(), JsValue> {
   wasm_logger::init(wasm_logger::Config::default());
